Feature/camera and UI (#6)
It adds more dynamic camera features to follow the player. (Thanks Quentin for the code!)

Adds a player UI.
Health bar that shows up when you are hurt with animation that fades in and out.
Health for the core.
Current charge of the core.
Timer for how long you have been playing.
Direction arrow to the center of the map where the core will be.